src-oxapy/pyproject.toml
[project]
name = "src-oxapy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-api-python-client>=2.186.0",
    "oxapy>=0.7.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy-libsql>=0.2.0",
    "yt-dlp>=2025.10.22",
]
